Allow assertions to ignore the beginning of a trace
Currently our flicker tests record only parts of the trace. The setup
and tear down actions on each test (e.g. start and app and close it to
test a rotation) are discarded.

For a better debugging it is interesting to allow traces to include more
information in the start and to ignore such failures. This allows the
trace to record more elements (test set up) which is important for
better understanding the actions during debugging.

Also, this implementation helps the tests to become more stable. When starting the tracing we record a "tracing.enable" record, which sometimes contains data which affects the test.

When no assertions pass it will still produce an error.

To use it set "skipUntilFirstAssertion" in the assertions.

Test: atest FlikcerTests
Change-Id: I9e296ae46664a2641e4bb91fea75729642e5c90a