Expand AbstractStandardAppHelper.
The goal is for ISettingsIntelligenceHelper to be able to open different
pages in Settings so that these pages can be benchmarked in Crystalball
tests.

In order to do this, we modified AbstractStandardAppHelper so that its
subclass can override getOpenAppIntent() which changes the the activity
intent used by open().

We also added a function in ISettingsIntelligenceHelper for the caller
to assign different actions wrt different Settings pages.

Test: forrest
Change-Id: I9137181e878d538feb57d4f4713bf4fe8e49b7f9