Add support to BufferStateLayer
BufferStateLayers are new layer types which are now drawn on screen.
This CL adds support for this layer type on the flicker tests.

To determie if a layer is visible we now check if it has child layers.
If not, we maintain the current visibility rules.

If the layer has children, we traverse the hierarchy until we reach a
leaf node and check if the leaf is visible instead.

Test: atest FlickerTests:ChangeAppRotationTest
Change-Id: Ie82cb79f83488177f27bc5775fb7e97712448fa9