Implementing lock contention stress mode
It's a InstrumentationRunListener that creates the lock contention stress while
the measured part of the test is running.

The parameters (intensity) of the stress are not yet finalized, but
having this stressor checked in will unblock things like creating
tests with enabled lock contention stress or experimenting with the stress
parameters.

Test: running tests with this stress mode
Bug: 168630376

Change-Id: I14098e878c5405b171d0795392bb8b32c2e1aabd