Adding skip for googlequicksearchbox crash
There is currently an intermittent failure in system_app_crash that is
affecting presubmit runs.  Under normal circumstances this skip would be
added after the crash has an owner.  In this case we know that the
instability is in AGSA on version 11.26 and it should be fixed once
version 11.30 is dropped.

As an interim, a very specific known failure skip is being added to
mitigate version 11.26.

Test: yes, forrest
Bug: 171971925
Change-Id: Ide6d84cf0b605875317649b55184fc8974df8338