Fixing filtering entries in ShownMoreThanOneConsecutiveEntry
For the assertions visibleWindowsShownMoreThanOneConsecutiveEntry and visibleLayersShownMoreThanOneConsecutiveEntry, for ignoring a list of entries, the check was assuming the layer.name / window.title would be inside the labels inside the input list, but it should be the other way around. In example, if we need to ignore an app layer, this will consist of the package name + activity name (i.e. my.package/my.package.AppActivity). To ignore the entry, it could be common that the person adding the assertion will add as an input "AppActivity" instead of the full package + activity name.

Also adding two tests to cover the two posible cases:
- One where we have the layer name that will be the same as the input layer (StatusBar#0 as layer name and ignored input).
- One where we have the layer name that will be larger than the input layer (com.google.android.apps.nexuslauncher/com.google.android.apps.nexuslauncher.NexusLauncherActivity#1 as the layer we need to ignore, but the input will be "Launcher").

Test: atest FlickerTests && atest FlickerLibTest
Bug: b/171049649
Change-Id: Id9b7026402a073461baebcdc94316cbd1f4df88a