Introduce Flicker- Subject, SubjectException and FailureStrategy
Currently the flicker subjects (Layers, WM and EventLog). This base class doesn't add any information related to the subject being tested (when using failWithoutActual) or the toString of the object (when using failWithActual).

Moreover, these methods are protected and cannot be accessed form the test specification. When a user throws a test exception the trace information is then missing.

With this FlickerSubject class we can ensure that the failures will always contain debugging information

Bug: 162923992
Test: atest FlickerLibTest
Change-Id: I33bb42284dff091f8063b51a45e4fe4d31f0238b