Provide the current state through WMStateHelper
This is necessary when tests need to obtain information (e.g. location) of windows or layers to interact with the current state.

For example, to interact with the PIP through UIAutomation, the most reliable way is to locate the PIP window and, obtain its coordinates and then interact with the location.

Bug: 172321238
Test: atest FlickerTests
Change-Id: Icb90c277576d279a29cb96ebf5732309268a1412