Simplify test factory + transition runner logic
Remove the factory cache and replace it for a cleanup method in the runner.

Instead of searching for a cached test, the test runner now receives a flicker object and an assertion to run and the flicker test rule is responsible for cleaning up the test.

This implementation allows the FlickerTestRule to be easily reused on different tests.

Bug: 178572522
Test: atest FlickerTests WMShellFlickerTests
Change-Id: I8c24894b58e8b54980d184b26a0c835142e3ccf5