Allow last/initial state assertion to access the whole trace
This is necessary to allow the final state to compare itself against a previous state in the trace.

For example, to detect that a pip window moved up when opening all apps, the most reliable metric is to identify the pip window position in the initial state and compare against its position in the last state.

Bug: 174238224
Test: atest WMShellFlickerTests:PipMovesInAllApps
Change-Id: Ifc7f70702ca4e83b6f3b0a5b611eb00a6ac0b2ac