Search for resumed activities in all leaf tasks
ON the WM hierarchy, tasks can be children of other tasks and only the leaf tasks have a `resumedActivity` value.
On flicker this information is used to ensure a state is complete or not.

Currently, flicker searches only the top root tasks to determine if they have an activity or not, instead it should look at all tasks to detect resumed activities in leaf tasks.

Moreover, it should also use the same logic for activities in displays and in tasks, instead of a different metric for each.

Test: atest FlickerLibTest
Bug: 167521849
Change-Id: I3e4824025fea54b7288ef99f58a0e57517afdb8f