Stabilize waitForRotation (wait for WM and SF states)
Currently the waitForRotation command waits only on the WM state. However, this may result in cases where the WM state is already correct but the SF state has not yet stabilized.

For this function to better reflect the device state it should wait on both the WM and SF states. Not it waits:
- for the specified rotation (WM state)
- until the rotation and black-surface layers are removed (SF state)
- until all layers have a simple rotation transform (SF state)

Test: atest FlickerTests
Bug: 181630910
Change-Id: I1c845e8d5636d935ec7ce482beeab731602d6547