Supports retrieving lists of layers or windowstates for assertions
In some tests it is necessary to compare the state of a layer or window against it's previous instance.

For example, on app launch it is necessary to check that the app window is expanding, that is, that it's visible region is always greater or equal to the visible region in the previous state.

This CL allows such behavior to be achieved using `WindowManagerTraceSubject.windowStates` and `LayersTraceSubject.layers`, both of which return a list of subject as they appear on the trace

Bug: 174238218
Bug: 174238219
Bug: 174238262
Bug: 174238224
Test: atest FlickerLibTest
Change-Id: I15be85f8d003c0ca4112aa6a9a14ff0e71dacda7