Remove applinktests
These tests check the verification state and the showing of the intent
disambiguation dialog. The former has been changed and replaced by
CtsDomainVerificationDevice/HostsTests and the latter has had its
behavior removed.

This test also relies on the YouTube package, which is arguably broken
behavior to begin with.

Will assume that other xTS tests will be enough to verify app links v2
behavior.

Bug: 179925390

Test: presubmit

Change-Id: I7d0449b2798fe0288f3025481159d6b3d6bd5dea