Enable recording Winscope trace files
Currently flicker automatically parses the trace objects when recording traces with `withTracing`

Sometimes, however, it is necessary to collect the proto files instead of the flicker objects (e.g. for future debugging or data gathering).

This CL makes the recording methods available (previously internal) and introduces a new method to obtain the trace files

Bug: 167521849
Test: atest FlickerLibTest
Change-Id: Ie8b432f9fa76b77a7af90ff6aeb3bb9924410097