Remove DeviceHealthTests
This test has been deprecated and replaced by DeviceHealthChecks a
long time ago, but it was never cleaned up. On top of that, the
crash checks in this test has been broken since Q release, leaving
confusions for users: they are not actually getting the coverage
of crash checks since all the crash checks will be reporting passing
regardless of actual failures.

Bug: 193459993
Change-Id: If5f94827c53f207fa48af097a0103d99932780f0
Test: compile, presubmit