Allow adding or subtracting regions for assertion
Shell transitions fix some underlying issues with layer calculations. Previously, when having a app (A) visible below a pip window, the visible region of (A) disregarded the existence of the pip layer (wrong).

Shell transitions fixes this issue and the visible region of (A) is correct, causing failures in the flicker tests

To simplify the test fixes, create a functionality to subtract the layer areas

Test: atest FlickerLibTest:LayersTraceSubjectTest
Bug: 193565745
Bug: 165793917
Change-Id: I20bb60a6e5ef6c93752d4621f18b2c7f121377f7
Merged-In: I20bb60a6e5ef6c93752d4621f18b2c7f121377f7