DO NOT MERGE: Disable FrameworksNetTests
Currently, there are some test for classes that are not in
connectivity module. If the platform code has a new change
that the test depends on it, then the test cannot verify the
behavior since the change does not exist in module branch.

This change moves the test for non-module classes out of
FrameworksNetTestsLib so that it doesn't go into coverage tests.
Then add the tests to FrameworksNetTests and disable the tests in
module branch.

Bug: 201265286
Test: make FrameworksNetTests fails with "unknown target" error
Change-Id: Iabb46726e92874e7fefde132b8ab9f102fac3cef