Generate bugreports for failed tests
Sometimes we have "system anomaly" popups
indicating that something went wrong at the system level,
such as ANRs:
https://android-build.googleplex.com/builds/I32100009991034417/git_master/7836278/test-mapping-presubmit/inv_16943106471426672245/TestScreenshot-BigPictureStyleNotification.bigPictureStyleNotificationTest_9686033669674506753.png

We take a bugreport at the end of the test suite,
but it may be too late because old diags may be lost.

Taking bugreport after each failed individual test is too expensive,
so we take bugreports only after the ones failed with an anomaly.

Bug: 202567877
Test: local, also prior knowledge from Launcher
Change-Id: Ib30177bba3664de5d9a7af1b254718dad24e8702