Opening full QS for the media carousel test
The test was failing because the test was opening
QQS instead of QS.

Fixing this by replicating
how OpenQsPanelAlone opens QS in 2 swipes.

Also fixing QuickSettingsHelper which
in some cases doesn't generate a long enough swipe.

Also, on devices with small screen, the carousel
is not visible after opening, and we need to scroll
up to make it visible.

All changes are non-elegant for now.
I replicate the existing type with hardcoded
constants and using timeouts. At some point I'll clean this
up.

Bug: 202567877
Test: atest PlatformScenarioTests:android.platform.test.scenario.sysui.media.QSMediaPlayerContent,android.platform.test.scenario.sysui.quicksettings.OpenQsPanelAlone
Change-Id: I1152d098a7f0c4da460cf24d0020bc5e0367880a