Patching CB tests incompatibility with TF
When the class-level part of the test fails (which includes an assumption exception used to dynamically skip tests), the test produces results that confuse TF. The result is messy output from atest (like something returned null) and messy Sponge pages, which include module-level errors, and a list of all errors listed for each of the tests:
https://screenshot.googleplex.com/BmRbaW6vkYw3q8U

As a patch, when the static part fails, we mimic that all test methods ran, and immediately failed as if they hit the same error.

There probably is a better solution, but it's beneficial to quickly fix the symptoms.

The outcome:
https://screenshot.googleplex.com/7p3rZdYkdqd5L9X

Bug: 203722419, 202567877
Test: local and lab runs
Change-Id: I53e7150b89e4af2579fa0bb9631090905c0bcb75