Support multiple components in the visible region calls
This is necessary to check, for example, that a specific set of layers covers the whole screen and it's used to enable the flicker tests in oriole.

In landscape mode, depending on app configuration, oriole devices include a `letterbox` layer where the notch is. To make the flicker tests work we need to consider both the notch and the app layers on the checks

Test: atest FlickerLibTest
Change-Id: I15e8394114c19a0b64223061b6d25fd92ee06d32