Crop layers to display size
Currently the flicker covers/occludes computation takes into account the entire layer bounds.

This causes layers to be identified as visible when their on screen area if fully covered, but their off screen area is not. Since the off screen area if not shown to the user the layer should be identified as invisible instead.

Bug: 207843938
Test: atest FlickerTests
Change-Id: I7cc5fa0fc6ddd8e874a9c9a85e926c1a59d994b8