Handle execution exceptions better
We want to make sure we handle and keep track of the different exection exceptions that appear so that we can better report them for easier debugging.

These changes insure that we have more granular details over which parts of the execution failed, while ensuring we still save the results of all succesfully executed transitions and run assertions on them.

Bug: 217196029
Test: atest FlickerLibTest:FlickerDSLTest
Change-Id: I4aa4616c69b4449a6c5ffa257b9999b43ae71030