Fix OpenCloseNotificationShadeInLockscreen test
As there was no Lockscreen Type set, `checkDeviceLock(true)` was failing. Calling goToLockScreen is enough in this case.

Then, the strategy to close the shade was not working on lockscreen (as pressBack doesn't work on lockscreen).
Now, the shade is closed with a swipe up, after reaching the bottom of it. This approach seems to work in all cases.

Fixes: 223360092
Test: atest PlatformScenarioTests:OpenCloseNotificationShadeInLockscreen#testOpenCloseNotificationShade -- --abi arm64-v8a
Change-Id: I5e0bdda744627d51afbfd7fc22ff7bc48db10218