Add a collector for bug report section durations
This change adds a collector and a helper that can extract a bug
report's section's durations from a bug report collected during a test.
Each section is returned as an individual metric.

The sections for showmaps have been excluded because they will vary
per-run (since the section names contain PIDs).

Test: - Ran CaptureBugReport with the new listener attached to confirm
        that the expected metrics were extracted from the bug report and
	returned.
      - Ran the included unit test.
Bug: 229396766
Change-Id: I519ea9886913ac3ac6af73b52d335de25f138b15