Add waitForValueToSettle
This is needed to reduce flakes related to getting a value. See cls in topic for usages.

The main reason for its introduction has been notification testing: if we just get the number of notifications visible on the screen after sysui has been restarted, this can fluctuate for a few seconds causing flakyness. With this, we're waiting for the value to settle for at least some seconds.

Bug: 226132311
Test: atest NotificationsVisibleOnLockscreen && atest NotificationsAndShelfVisibleOnLockscreen
Change-Id: Ic3c4b68fa035f886246db60c2f1c210c4adbed21