Add top gesture margin to apps list in launcher to fix scroll up
During making a massive amount of start-up measurements on my Wembley found out that sometimes
tests are failed with:

android.support.test.uiautomator.StaleObjectException
 ...
 at android.support.test.launcherhelper.CommonLauncherHelper.launchApp(CommonLauncherHelper.java:134)
 at android.support.test.launcherhelper.BaseLauncher3Strategy.launch(BaseLauncher3Strategy.java:226)
 at android.platform.helpers.AbstractStandardAppHelper.open(AbstractStandardAppHelper.java:156)
 ...

After writing a stress test opening the apps screen and trying to scroll it to the top:

while (true) {
    UiObject2 container = launcherStrategy.openAllApps(false);
    scrollBackToBeginning(container, Direction.reverse(dir));
}

was able to easily reproduce the problem. The main reason of the problem is the fact that
the search bar overlaps the apps recycler view a little bit, so because the scroll is trying to start
swipes from the very top, instead of scrolling the list up we're just closing apps page. Usually
swipe is not long enough to close the screen, but in rare cases, it actually makes the thing and we end up
with the exception I mentioned above. Video: https://drive.google.com/file/d/1U4UrG23q5EX92gqfat_L3ispqgQAiC49/view?usp=sharing

Solving this by adding a gesture margin for about 15% of the whole app's recycler view heigh will allow us to actually swipe the recycler view instead of closing the apps page.

Test: Stress test after applying a fix works fine
Change-Id: I4ae54ea343f8c207c0ad3fc28d76a0f076ad4e8a