Report Flicker execution errors from Runner
This ensures that Flicker execution errors are always reported to the
notifier which means they show up as a test failure and is immediately
visible that an error occured rather than just appearing in the logs but
with a green test (in the case that at least 1/x transitions executed
successfully).

This also fixes the problem with the results not being cleaned up
properly leading to OOM excpetion if we had a execution exceptions but
at least one successful transition run.

Bug: 236592450
Bug: 236687168
Test: atest FlickerLibTest

Change-Id: Ie6255138bd278204e25355687ac48ac1ad27a306