Introduce the DeviceEmulationRule for screenshot tests
This CL introduces a new DeviceEmulationRule, forked from SystemUI's
ScreenshotTestRule [1], which allows to emulate a device configuration
(display, theme and orientation) on any device. This allows to test
multiple different configurations on a single device very easily,
without having to own the emulated device. In particular, this allows to
test phone, tablet and foldable configurations on Cuttlefish.

Tests using this rule should make sure that the UI under test is *not*
device-specific, otherwise the test will have different behaviors
depending on which device the test is run. This makes this rule
especially useful for *unit* screenshot tests, but it could still be
used for integration screenshot tests when applicable. A big advantage
of forcing the screenshot tests using this rule to be device-agnostic is
that it sets up for the future, once host (device-less) screenshot tests
will be available.

An example test with device emulation is added in ag/19167907.

[1] http://cs/android-internal/frameworks/base/packages/SystemUI/screenshot/src/com/android/systemui/testing/screenshot/ScreenshotTestRule.kt

Bug: 230832101
Test: atest DeviceEmulationRuleTest
Change-Id: I1d8decb9b3fa75c32346c852ac30e77c45f8a132