Add golden identifier to PNGs saved on device
This CL adds the golden identifier of the current screenshot test when
saving PNG files to the device under test. This will allow pulling the
files for a screenshot test when emulating a device very easily, which
is convenient when working locally and when running tests with Android
Studio/Gradle instead of Soong (like SystemUI does).

Bug: 230832101
Test: atest SystemUIGoogleScreenshotTests
Change-Id: I8afdfb076f4cbd480ac7409f8325e86cea810ee1