Force touch mode for screenshot tests
Enables touch mode before running screenshot
tests because in touch mode Android automatically
requests focus for the first element on the screen.
This might change the look of some elements
on the screen.

Previously, we added clearFocus() call but this
might not be enough in some cases when the app under
test reacts to the focus in non-reversible way
(e.g. expands bottom sheet).

Test: run screenshot test when in keyboard mode
 => check that it resets it and captures a screenshot
    in touch mode (without foucsed elements)
Bug: 245354130
Change-Id: I7c92af3120fa01d7913f1167bc1828f543c1dd4c