Enabled all assertions on FaaS and type matcher fix.
Set to use all assertions (both hardcoded and generated) in Flicker as a
Service.
Added minor fixes in response to integration CL comments: https://googleplex-android-review.git.corp.google.com/c/platform/platform_testing/+/20082150
Fixed bug that was not generating assertions for non-hardcoded
components because of unflattenFromString function -> now the key in the
trace lifecycle can be directly the type component.

Test: atest FlickerLibTest:AssertionsTest
Bug: 241905004
Change-Id: Id67338b99fba701d62bdc0cc6e4bb8fa3d5506e3