Refactor HealthTestingUtils to use WaitUtils
This change updates the implementation of HealthTestingUtils to use WaitUtils under the hood, while maintaining the same public interface. This will provide better logging and allow us to see performance data in perfetto.

There is one minor behavior change to be aware of: the previous implementation of HeathTestingUtils.waitForCondition would perform a final check after the timeout had expired, while the new version does not. We will monitor perfetto traces of failed tests to determine if this late check was necessary.

Bug: 262588714
Test: presubmits
Change-Id: Ieec2ac50e64e7e6f8e7998b21e1e5ebb92363d93