WaitCondition: more retries, shorter timeouts
There are a lot of places where WaitCondition waits for one timeout,
then succeeds. That means the default timeout is too long and is
wasting time. Moving to 50 retries after 100ms seems like a ~10%
improvement in some of the longer EnterPipTest subtests.

That said, a dedicated event trigger is still better than a timeout.

Test: atest WMShellFlickerTests:EnterPipTest
Bug: 112837428
Change-Id: Ib168d0c291445500e9eda535770cd9e4329b4c05