Sleep after folding in tests
The fold behaviour has recently been changed to keep displays on if there is a wakelock held. This causes non-deterministic tests behaviours.

This adds a sleep after folding, to make the fold behaviour deterministic.

As we don't need it for now, this sleep is not configurable (e.g it always happens after folding)

Test: UnfoldToLauncher
Bug: 266207627
Change-Id: I0d966a549090095d4ed2bd10d148e0c151fc49f1