Split component id matcher into strict and relaxed matchers
So that we can match an entire component group by id as well instead of just a specific layer. This is useful for FaaS where we get the task layer id for example and want to check the task has visible activity layers in it. In this case we want to match the task and all of its children since the task is just a container so not visible.

Bug: 265791451
Test: atest FlickerLibTest
Change-Id: Ibc5c0b2da030973d65a1978aa2897dab15bba266