Make use of newly dumped finish time of transitions
We were relying on using the finish transaction id to figure out when a transition ended, but that isn't reliable if the transition runner decides to merge this transition with another transition leading to a transition with a different id being applied that we cannot track. So instead we need to dump the finish time of a transition. A side effect is we will no longer dump a transition when it is sent to shell but only once it is finished or aborted.

Fixes: 269279367
Test: atest FlickerLibTest
Change-Id: Ie3e376d1a1aadfe737ee33d97d6a2496090f1404