Disable FaaS scenarios that are not validated
The scenarios being disabled are not working in all cases. Notable many of these scenarios do not trigger WM or SF changes which leads to errors when reading the WM or SF traces that contain 0 or 1 entry. Until we properly handles these scenario they should be disabled to avoid test failures.

Bug: 270949117
Test: atest FlickerLibTest
Change-Id: I9327a76f4daab39e085be0778cf235b6bd133520