Clear output dir on test run start rather than trace collection
When using the FlickerServiceRule we will likely be running lots of tests and collecting traces for all those tests and we don't want to delete all the artifacts before we start collecting traces for every test. We only want to remove the files onces all tests in a test run have executed and CB has pulled the trace files off the device.

Bug: 270949117
Test: atest FlickerLibTest
Change-Id: I8e443af9834aacec8131b3fc2ef2995d943bd838