Add unique group ID to notifications
The notifications scroll jank test is supposed to create 5 custom
notifications for test purposes and scroll through them. Lately we've
been seeing that these notifications are being grouped into one. This
grouping reduces the number of notifications available in the panel and
subsequently fails the condition required for being able to "scroll"
through them. This needs to be fixed to ensure that each notification
added shows up as a separate notification on the notifications panel.

Test: make catbox
This has been locally tested by executing the test on a reference device

Bug: 274826074
Change-Id: I3bf7a73bf1c1b24979ceb19adb5e8e2455bc0c59