Update helper to ignore wifi/telephony bug reports
This change updates BugReportDurationHelper to filter out wifi/telephony
bug reports, because they break the assumption that lexicographically
greater bug report names are from more-recent timestamps. This change
ensures that the bug report being inspected for section durations is the
one captured during the CaptureBugReport test, and not a wifi or
telephony bug report that was taken before the test.

This change also:
- adds additional logging for bug report names and unzip commands being
  run during the test
- dumps all bug report entries if a bug report archive is found, but
  nothing was read from it

Test: Run health/bugreport/capturebugreport on bluejay and verify using
      the new logs that the bug report being inspected for section
      durations is the latest bug report (i.e. a wifi/telephony bug
      report isn't being opened).
Bug: 277202149
Change-Id: I9beced453bf672534695801504094d04c9fe02d2