Add hot and cold start up specific JUnit rules
The idea here is to separate the app start up run rule into two separate
rules with the prime objective of being able to pass different values of
the same parameter based on the scenario (hot or cold start up).

Test: make catbox
This change has been tested manually by executing the test on a locally
connected reference device.

Bug: 278622162
Change-Id: I60c7abd13e8eaa2daeafc189cc3f04b26f3e6f3c