Migrate some SysUI test rules to shared component
This was per gelanchezhian@'s request so that the rule files don't
have to be duplicated when we set up the new package in ag/22514688.
The `TestIsolationRule` was left in place (to be duplicated in the
new package) since it has dependencies on `systemui_tapl` that can't
yet be migrated to AOSP.

Bug: 268530721
Test: Combined with ag/22720241, by invoking:
atest
PlatformScenarioTests:android.platform.test.scenario.sysui.sharesheet.ShareScreenshot
Also confirmed that this failed to build while the migration was
in-progress, in order to verify that the above atest invocation in
fact exercises the code being modified in these CLs.

Change-Id: I9e5446ae6ffea572eb8750397459dd4abe9e555f