Do not upload passing tests' archifacts.
As an effort to work around b/283089903, this CL stops uploading
archifacts and screenshots of passing CLs. As commented by Jordan in
https://buganizer.corp.google.com/issues/283089903#comment6, the
archifacts of previous passing tests blocked the uploading of following
failing tests if their files share names. This CL makes sure the
uploaded archifacts/screenshots correspond to a failing test so that
they are at least useful.

The ideal solution would be including the orientation (i.e. landscape
vs. portrait) in the archifacts' names (so that the two tests could be
distinguished). We need to investigate further for including this
information to the archifacts.

Bug: 283089903
Test: Unit tests/Presubmit.
Change-Id: I7a255209836546422666e605b02bb7f3aee64613