Migrate typing logic from UiBench app to UiBenchMicrobenchmark test
This CL is a follow-up CL for [1] that we found in PTS build environmant
cannot use a generic platform key to sign the apk accorss devices for
testing (but using atest could pass the test), so the UiBench app still
not able to use injectInputEvent because of lacking INJECT_EVENTS
permission.

As a result, we realized that injectInputEvent logic could migrate to
the instrument test process which makes more sense as the remote test,
so that instrument process could leverage {adopt, drop}ShellIdentity to
pass the permission check without worrying about the apk signing.

Also add a receiver to for test side to cancel the typing if the test
activity was paused.

Also, add a device configuration in unibench-all.xml to ensure
`test_harness=1` property will set automatically when using atest to
trigger the test.

[1]: Id3c3490d7ab98a17da4e07d8d62d048c98ff1ec4

Bug: 289210988
Test: atest UiBenchEditTextTypingMicrobenchmark
Test: pts-tradefed run commandAndExit pts-postsubmit-very-long -m PtsUiBench \
  --fastboot-command-preparer:disable  \
  --log-level-display verbose -l VERBOSE \
  --test-arg com.android.tradefed.testtype.AndroidJUnitTest:instrumentation-arg:iterations:=1 \
  --test-arg com.android.tradefed.testtype.AndroidJUnitTest:class:\
com.android.uibench.microbenchmark.UiBenchEditTextTypingMicrobenchmark#testEditTextTyping

Change-Id: I443f67748dfd3db0deb4cd12ac068975fbebc2c3