Group all assertions in a doEvaluate into one
Instead of having an assertion for every single call to flicker in an AssertionTemplate's doEvaluate group all those assertions into one assertion result.

Also include any assertion thrown outside of flicker blocks in the assertion result. This allows you to write assertions outside of the context of the Flicker objects for more flexiblity.

Test: atest FlickerLibTest
Change-Id: Ia7b66f109ce56b131168ae3d4c0028578628ad8e