Enable SystemUIGoogleScreenshotTests on main and disable on udc-qpr-dev
At the moment the tests are only running on `udc-qpr-dev`, and not
on `main`.
If we disable them on `udc-qpr-dev` and keep them enabled on `main`,
they will still run in presubmits of both branches.

This way we also solve the problem of having potentially different
golden differences in both branches.

Fixes: 292206289
Change-Id: I45b5f7bdf8f8346846d12f455c56a41e20636443