Invoke UserManager API using Shell permissions
Function calls involving UserManager API need to be invoked with
explicit CREATE_USERS permissions. Failing to do so results in a
java.lang.SecurityException.

The function getAliveUsers() has been replaced with getUsers() to fix a
java.lang.NoSuchMethodError.

Test: make ats
This change has been tested locally on a reference device flashed with
the following -
 - Android T
 - Android U

The following test plans were run successfully with the above builds -
 - catbox-performance-switch-to-existing-user
 - catbox-performance-create-and-switch-to-new-guest

Bug: 279227985
Change-Id: I97e7858a46f4e9dccc109ab3f6faf130cec6646a