Make MaterialYouColorsRule work with new Material colors (1/2)
The previous implementation of MaterialYouColorsRule was not working
with the new material you tokens. This CL change the implementation of
this rule to override the setting used to compute the colors, like we do
in MaterialYouPaletteTest.

Given that this CL will require updating all existing golden images
using MaterialYou colors, I changed the fixed colors from the previous
ugly orange to the nice green/blue presets available in the wallpaper
picker.

Bug: 290209259
Test: atest MaterialYouColorsRuleTest
Test: atest SystemUIGoogleScreenshotTests
Change-Id: I06db11a041bac5daaa9281b4ce4838c638901d53