Modify the logic to generate artifact's name.
The purpose of this CL is to include golden image filename to the
(temporary) artifact's name (e.g. the actual screenshot name and the
proto filename). Since the golden image filename includes
device config information, it will deduplicate the filename stated in
the bug 296418877.
ag/24472563 also serves for this purpose, but did it not completely: it
only changes the filename of all screenshots.

Bug: 296418877
Test: atest -c SystemUIGoogleRoboRNGTests:com.android.keyguard.KeyguardClockSwitchScreenshotTest
    (http://ab/I62200010188571652, click `Update scuba goldens` and we
    could see that there are in total eight screenshots.)
Change-Id: I0fd3e65be1933ce5e5b06758066459d87a55ee67