Remove logcat buffer size dependency in LogcatInspector
Currently, LogcatInspector looks for all to-be-matched logs sequentially in a single logcat buffer. This is done to ensure that the logs are present in chronological order.

Backup&Restore xTS tests have been flaky due to this buffer size limitation. This has previously been mitigated by increasing the logcat buffer size to it's max limit of 16M (ag/20271867). But there are some devices, like Go devices, for which the max logcat buffer size is 1M.

To remove the dependency on logcat buffer size, this change looks at the timestamp of log and compares that with the next matched timestamp, which ensures that logs are present in chronological order.

Bug: 295483060
Test: Run Backup&Restore CTS and GTS tests
Change-Id: Icf7084fe5509045737b408296e9c46c557d03906