Change FlagSetException to unchecked exception
FlagSetException should be thrown when the flag cannot be set to
ceratain value successfully. If in a test gets the FlagSetException, it
means the test wants to set the value of the flag, but failed to do it.
The test should not be expected to catch this failure and handle it,
since it cannot really handle it.

Test: atest FlagJUnitHostTests --host
Bug: 295561701
Change-Id: Ib494e001dc557ca5e76f7042bc41cdb9b1b0a2be