SetFlagsRule reset FeatureFlag in Flags to FeatureFlagsImpl in reset
Before once a flag is managed by SetFlagsRule, then after each test
method, SetFlagsRule will set the flag value to null, and keep the Flags
class using the FakeFeatureFlagsImpl.

This behavior causes a problem. If one flag is managed by SetFlagsRule
in one test class, then other test class even this flag is not intended
to be set, then the user has to set the value of the flag, since the
implementation has been switched in the Flags class.

Bug: 299360354
Test: atest FlagJUnitHostTests --host
Change-Id: I3209e3098c11fc79bd46b2c56091d78ba8194e8a