Add multi user test skip condition in test set up
The multi user tests build with AOSP target are supported on devices
running Android T or higher. This condition has been added to the set up
section of the test. Additionally, a test module configuration is being
added so that specific module lever target preparers can be executed
before the test module. This is necessary so that these tests can be
integrated into a consolidated test plan in the future.

Test: make catbox
These changes have been tested on a locally connected reference after
flashing the device with Android S and Android T

Bug: 302028453
Change-Id: If6926dedd04ee51cb849dbd56e1f1c4c2f765644