Fix display size cache
Previously we stored up to 1 item, which got cleared whenever there was a rotation or the display went off. Since the objects are small (rect), store the last 4 display sizes to mitigate the number of times we have to dump the device state.

On local tests this reduces the assertion runtime of `ShowPipAndRotateDisplay` from 3s to 2ms

Bug: 301244689
Bug: 301574268
Bug: 301574153
Test: atest FlickerLibUtilsTest
Change-Id: Ibc12aea6e1bf530ce0b1670ea970e321c4185eb7