Look for clock container to find lockscreen
Right now, we use the lockscreen clock (or clock_large) as a
representative view for the lockscreen -- if it's visible, then the
we know the lockscreen is.

Unfortunately, since we added KeyguardClockSwitch, it's possible for the
lockscreen clock to switch out from under code that's waiting for it to
become visible, resulting in a StaleObjectException in methods like
LockscreenUtils.ensureLockscreen.

This change uses the clock *container* as a representative view instead.
A quick look at the code suggests that the clock container is never
hidden, just the clock frames themselves, so this should reduce the risk
of spurious StaleObjectExceptions.

Bug: 308234245
Test: atest LockscreenNotificationsOverlapWithLockIcon
Flag: NA
Change-Id: I04280add56a8a617ee134c398abd6a54eb9bec78